Add component test cases
